pyproject.toml
[project]
name = "my-repos-dashboard"
version = "0.1.0"
description = "Local Git repository and worktree management dashboard"
readme = "README.md"
authors = [
    { name = "MohamedHamed19m", email = "mohamedahamed1915@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
my-repos-dashboard = "my_repos_dashboard.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["pytest>=7.0.0", "httpx>=0.24.0"]
